refactor(ai): relocate AI services from Core to AI project
Move 7 AI service files + 7 value types/interfaces from Core/Services/AI/
to AI/Services/. Namespace flip from TelegramGroupsAdmin.Core.Services.AI
to TelegramGroupsAdmin.AI.Services. AI's ServiceCollectionExtensions now
owns the four AI registrations; Core's drops them. Microsoft.SemanticKernel
package moves from Core's csproj to AI's. Telegram csproj gains a
ProjectReference to AI.

Build is intentionally broken at this commit — consumer using-statement
updates land in commit 6 alongside the rest of the consumer sweep.

Refs spec: docs/superpowers/specs/2026-04-25-config-and-ai-relocation-design.md
