+++ TelegramGroupsAdmin.AI/Services/AIServiceFactory.cs
@@ -3,7 +3,7 @@ using Microsoft.Extensions.Logging;
 using TelegramGroupsAdmin.Configuration.Models;
 using TelegramGroupsAdmin.Configuration.Repositories;
 
-namespace TelegramGroupsAdmin.Core.Services.AI;
+namespace TelegramGroupsAdmin.AI.Services;
 
 /// <summary>
 /// Service for AI configuration management (used by Settings UI)
