+++ TelegramGroupsAdmin.AI/Services/ChatCompletionOptions.cs
@@ -1,4 +1,4 @@
-namespace TelegramGroupsAdmin.Core.Services.AI;
+namespace TelegramGroupsAdmin.AI.Services;
 
 /// <summary>
 /// Options for chat completion requests
