+++ TelegramGroupsAdmin.AI/Services/ChatCompletionResult.cs
@@ -1,4 +1,4 @@
-namespace TelegramGroupsAdmin.Core.Services.AI;
+namespace TelegramGroupsAdmin.AI.Services;
 
 /// <summary>
 /// Result from a chat completion request
