+++ TelegramGroupsAdmin.AI/Services/IFeatureTestService.cs
@@ -1,6 +1,6 @@
 using TelegramGroupsAdmin.Configuration.Models;
 
-namespace TelegramGroupsAdmin.Core.Services.AI;
+namespace TelegramGroupsAdmin.AI.Services;
 
 /// <summary>
 /// Service for testing AI feature configurations before saving
