+++ TelegramGroupsAdmin.Telegram/TelegramGroupsAdmin.Telegram.csproj
@@ -28,6 +28,7 @@
   <ItemGroup>
     <ProjectReference Include="..\TelegramGroupsAdmin.Configuration\TelegramGroupsAdmin.Configuration.csproj" />
     <ProjectReference Include="..\TelegramGroupsAdmin.Data\TelegramGroupsAdmin.Data.csproj" />
+    <ProjectReference Include="..\TelegramGroupsAdmin.AI\TelegramGroupsAdmin.AI.csproj" />
     <ProjectReference Include="..\TelegramGroupsAdmin.ContentDetection\TelegramGroupsAdmin.ContentDetection.csproj" />
     <ProjectReference Include="..\TelegramGroupsAdmin.Core\TelegramGroupsAdmin.Core.csproj" />
   </ItemGroup>
